requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Flask-Cors
Flask-RESTful
Flask-JWT-Extended
flask-admin
requests
numpy
python-dateutil
tzdata
tzlocal
pytz
regex
six
certifi
idna
colorama
click
MarkupSafe
Werkzeug
dateparser
openai
cryptography
passlib
transformers
pyyaml
dateparser
python-dotenv  # Environment variable management for development
psycopg2-binary  # PostgreSQL database adapter for Python